+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
  - Add way to get users for workshops
  - Added functionality search beneficiaries
+ - Added functionality view a Workshop (Video functionality needs more info)
 
 ## [0.1.0] - 15/11/2025
 ### Fixed